+++ src/skills.js
@@ -1,0 +1,3 @@
+export const skills = [
+    "React JS", "Solidity", "Python", "Polygon"
+];
